add: basic lazygit install
