--- lazygit_rocker/templates/new_rocker_extension_snippet.Dockerfile
@@ -1,2 +1,0 @@
-RUN curl -fsSL https://pixi.sh/install.sh | bash
-RUN echo 'eval "$(pixi completion --shell bash)"' >> ~/.bashrc
