+++ test/test_basic.py
@@ -2,26 +2,26 @@
 import argparse
 import pytest
 from unittest.mock import patch
-from lazygit_rocker.new_rocker_extension import NewRockerExtension
+from lazygit_rocker.lazygit_rocker import LazygitExtension
 
 
 class TestNewRockerExtension:
 
     # Instantiating PixiExtension and verifying the name attribute is set correctly
     def test_name_attribute_initialization(self):
-        extension = NewRockerExtension()
+        extension = LazygitExtension()
         assert extension.name == "new_rocker_extension"
 
     def test_register_arguments(self):
         parser = argparse.ArgumentParser()
-        NewRockerExtension.register_arguments(parser)
+        LazygitExtension.register_arguments(parser)
         args = parser.parse_args([])
         assert "new_rocker_extension" in vars(args)
 
     # Handling missing template files in get_snippet method
     def test_get_snippet_missing_template(self):
 
-        extension = NewRockerExtension()
+        extension = LazygitExtension()
         with patch("pkgutil.get_data", return_value=None):
             with pytest.raises(AttributeError):
                 extension.get_snippet({})
@@ -29,7 +29,7 @@ class TestNewRockerExtension:
     # Retrieving the default snippet using get_snippet method
     def test_retrieve_default_snippet(self):
 
-        extension = NewRockerExtension()
+        extension = LazygitExtension()
         snippet = extension.get_snippet(None)
 
         assert snippet is not None
@@ -37,7 +37,7 @@ class TestNewRockerExtension:
     # Retrieving the user-specific snippet using get_user_snippet method
     def test_retrieve_user_specific_snippet(self):
 
-        extension = NewRockerExtension()
+        extension = LazygitExtension()
         snippet = extension.get_user_snippet(None)
 
         assert snippet is not None
